changed according to comments and moved paginatedSyncApi call to HomeFeed
- moved paginatedSyncAPI call to homeFeed to maintain parity between dmFeed and groupFeed as we are using common timestamp for both
